requirements.txt
asgiref==3.3.4
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.7
defusedxml==0.7.1
Django==2.2
django-allauth==0.40.0
django-debug-toolbar==1.9.1
django-jsonfield==1.3.1
django-markdown-filter==0.0.1
django-markdownify==0.8.0
django-mdeditor==0.1.18
django-pandas==0.6.4
django-six==1.0.4
django-taggit==1.4.0
django-utils-six==2.0
idna==2.10
importlib-metadata==4.0.0
ipython==5.8.0
ipython-genutils==0.2.0
Markdown==3.3.4
markdown2==2.4.0
mistune==0.8.4
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==1.0.18
ptyprocess==0.7.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
simplegeneric==0.8.1
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.1
#
django-bower
psycopg2-binary>=2.8
gunicorn==20.0.4
# recupere les vars declarer dans le fichier .env.prod
python-dotenv
